Update javascript-methods.js
map()
Without using the native "Array.prototype.map" method of JavaScript, compose a function titled "myMap" that will take in an array of elements and execute a callback function on each of those elements. Syntax: myMap(callbackFn)
filter()
Without using the native “Array.prototype.filter” method of JavaScript, compose a function titled "myFilter" that will take in an array of elements and execute a callback function on each of those elements. Syntax: myFilter(callbackFn)
some() 
Without using the native "Array.prototype.some" method of JavaScript, compose a function titled "mySome" that will take in an array of elements and execute a callback function on each of those elements. Syntax: mySome(callbackFn)
every()
Without using the native “Array.prototype.every” method of JavaScript, compose a function titled "myEvery" that will take in an array of elements and execute a callback function on each of those elements. Syntax: myEvery(callbackFn)
reduce()
Without using the native "Array.prototype.reduce" method of JavaScript, compose a function titled "myReduce" that will take in an array of elements and execute a callback function on each of those elements. Syntax: myReduce(callbackFn)
includes()
Without using the native "Array.prototype.includes" method of JavaScript, compose a function titled "myIncludes" that will take in an array of elements and indicate whether or not a target element is contained within the input array. This returns a boolean. Syntax: myIncludes(searchElement)
indexOf()
Without using the native "Array.prototype.indexOf" method of JavaScript, compose a function titled "myIndexOf" that will take in an array of elements and returns the index of the first encounter of a target element (if it is found) or -1 if that element does not exist within the input array. Syntax: myIndexOf(searchElement)
lastIndexOf()
Without using the native "Array.prototype.lastIndexOf" method of JavaScript, compose a function titled "myLastIndexOf" that will take in an array of elements and returns the index of the last encounter of a target element (if it is found) or -1 if that element does not exist within the input array. Syntax: myLastIndexOf(searchElement)
Object.keys()
Without using the native "Object.keys" method of JavaScript, compose a function titled " Object.myKeys" that will take in an object and return all of the keys of the key:value pairs of that object. Syntax: Object.myKeys(object)
Object.values()
Without using the native "Object.values" method of JavaScript, compose a function titled " Object.myValues" that will take in an object and return all of the values of the key:value pairs of that object. Syntax: Object.myValues(object)